+++ composeApp/src/commonMain/kotlin/viewmodels/EditBucketComponent.kt
@@ -7,16 +7,17 @@ import com.technology626.budgyt.budgyt
 import models.Bucket
 import models.BucketType
 import java.util.UUID
+import java.math.BigDecimal
 
 interface EditBucketComponent {
     val bucket: Bucket?
-    fun addBucket(bucketName: String, bucketType: BucketType, bucketEstimate: Float)
+    fun addBucket(bucketName: String, bucketType: BucketType, bucketEstimate: BigDecimal)
 
     fun editBucket(
         bucketId: UUID,
         bucketName: String,
         bucketType: BucketType,
-        bucketEstimate: Float
+        bucketEstimate: BigDecimal
     )
 }
 
@@ -26,13 +27,13 @@ class DefaultEditBucketComponent(
     private val database: budgyt,
     val onAddBucket: (bucketId: UUID) -> Unit
 ) : EditBucketComponent, ComponentContext by componentContext {
-    override fun addBucket(bucketName: String, bucketType: BucketType, bucketEstimate: Float) {
+    override fun addBucket(bucketName: String, bucketType: BucketType, bucketEstimate: BigDecimal) {
         val newBucketId = UUID.randomUUID()
         database.bucketQueries.addBucket(
             id = newBucketId,
             bucket_name = bucketName,
             bucket_type = bucketType,
-            bucket_estimate = bucketEstimate.toDouble()
+            bucket_estimate = bucketEstimate
         )
         onAddBucket(newBucketId)
     }
@@ -41,12 +42,12 @@ class DefaultEditBucketComponent(
         bucketId: UUID,
         bucketName: String,
         bucketType: BucketType,
-        bucketEstimate: Float
+        bucketEstimate: BigDecimal
     ) {
         database.bucketQueries.editBucket(
             bucket_name = bucketName,
             bucket_type = bucketType,
-            bucket_estimate = bucketEstimate.toDouble(),
+            bucket_estimate = bucketEstimate,
             id = bucketId
         )
         onAddBucket(bucketId)
